% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPhi.R
\name{bootPhi}
\alias{bootPhi}
\title{Properties of bootstrap confidence intervals for the kinship coefficient}
\usage{
bootPhi(
  ped,
  ids = NULL,
  N = 2,
  B = 2,
  CItype = "bca",
  conf.level = 0.95,
  plot = F,
  seed = NULL,
  verbose = F
)
}
\arguments{
\item{ped}{ped object with allele frequencies.}

\item{ids}{Id of pair.}

\item{N}{Integer. No of simulations.}

\item{B}{Integer. No of bootstraps.}

\item{CItype}{Logical. See \code{\link[=interval]{interval()}}}

\item{conf.level}{Double}

\item{plot}{Logical}

\item{seed}{Integer.}

\item{verbose}{Logical.}
}
\value{
A list with the following four elements:
\itemize{
\item \code{phi}: The kinship coefficient of the pedigree.
\item \code{averaged} : A data frame with two lines, one for \code{parametric}
and one for \code{nonParametric}. The values are averaged over the \code{N} simulations.
The columns \code{realised}, \code{boot}, \code{skew}, \code{lower},  \code{upper}, and  \code{cover} are explained
in \code{Details.}
\item \code{simParametric}. The entries are as for \code{averaged}
but for each parametric simulation.
\item \code{simNonparametric} The entries are as for \code{averaged}
but for each nonparametric simulation.
}
}
\description{
The purpose is to compare parametric and nonparametric confidence
intervals in kinship applications, currently only for the kinship coefficient.
}
\details{
Marker data are simulated \code{N} times giving \code{N} estimates of (kappa0, kappa1, kappa2).
For each simulation, the \code{realised} phi is found. Parametric and nonparametric bootstrapping
is done giving \code{boot} (the averaged kinship coefficient from \code{B} bootstrap simulations)
and \code{skew} = \code{realised}- \code{boot}. There are various
ways to calculate the confidence interval. The default is \code{bca} as implemented in \code{coxed::bca}. The
variable \code{cover} is 1 if the confidence interval contains \code{phi} and 0 otherwise.
}
\examples{
# The example considers the kinship coefficients between
# brothers named `B1` and `B2` using the 35 markers in
# forrel::NorwegianFrequencies.

library(forrel)
library(ribd)
library(coxed) # for bca confidence intervals

ids = c("B1", "B2")
ped = nuclearPed(2, children = ids)
ped = setMarkers(ped, locusAttributes = NorwegianFrequencies)
N = 10 # no of confidence intervals. Increase
B = 100 # no of  bootstraps. Increase
res1 = bootPhi(ped, ids, N, B, seed = 17)

# Basic output
res1[1:2]

# Compare parametric and nonparametric estimates
y1 = res1$simParametric$boot
y2 = res1$simNonparametric$boot
boxplot(y1, y2, names = c("parametric", "nonparametric"),
        main = "Bootstrap estimates of kinship coefficient",
        sub = "Red stapled line: theoretical value")
abline(h = res1$phi, col = 'red', lty = 2)
}
